6757: Selector for attribute in the flame graph
Add support for grouping Flamegraph View and Stacktrace View by
custom numeric event attribute.
For example instead of grouping stacktraces by number of samples for
the stacktrace, we group by Allocation Size, which can give us places
where we allocating the largest buffers.

+++ application/org.openjdk.jmc.flightrecorder.ui/src/main/java/org/openjdk/jmc/flightrecorder/ui/pages/MethodProfilingPage.java
@@ -160,7 +160,7 @@ public class MethodProfilingPage extends AbstractDataPage {
 			if (event.index == 2 && (total = rootFork.getItemsInFork()) > 0) { // index == 2 => percentage column
 				// Draw siblings
 				Fork parentFork = frame.getBranch().getParentFork();
-				int forkOffset = parentFork.getItemOffset();
+				long forkOffset = parentFork.getItemOffset();
 				int siblingsStart = (int) Math.floor(event.width * forkOffset / total);
 				int siblingsWidth = (int) Math.round(event.width * parentFork.getItemsInFork() / total);
 				event.gc.setBackground(SIBLINGS_COUNT_COLOR);
